\name{alt.at}
\alias{alt.at}
\docType{data}
\title{
Altitude data from Austria
}
\description{
a \eqn{585 \times 314}{585 x 314} matrix with altitude data representing a grid covering Austria
}
\seealso{
\file{\link{lon.at}}, \file{\link{lat.at}}
}
\usage{data("alt.at")}
\examples{
alt = get(data("alt.at"))
}
\keyword{datasets}
