\name{lat.at}
\alias{lat.at}
\docType{data}
\title{
Latitude data from Austria
}
\description{
a \eqn{585 \times 314}{585 x 314} matrix with latitude data representing a grid covering Austria
}
\usage{data("lat.at")}
\seealso{
\file{\link{lon.at}}, \file{\link{alt.at}}
}
\examples{
lat = get(data("lat.at"))
}
\keyword{datasets}
