\name{lon.at}
\alias{lon.at}
\docType{data}
\title{
Longitude data from Austria
}
\description{
a \eqn{585 \times 314}{585 x 314} matrix with longitude data representing a grid covering Austria
}
\usage{data("lon.at")}
\seealso{
\file{\link{lat.at}}, \file{\link{alt.at}}
}
\examples{
lon = get(data("lon.at"))
}
\keyword{datasets}
